separate rust code into project files and try built-in typed-arena for quadtree nodes
