+++ polytope_feature/datacube/quadtree/Cargo.toml
@@ -8,6 +8,7 @@ pyo3 = { version = "0.20", features = ["extension-module"] }
 geo = { version = "0.30"}
 qhull = { git = "https://github.com/LucaCiucci/qhull-rs" }
 rayon = "1.8"
+typed-arena = "2.0"
 
 [lib]
 name = "quadtree"
